+++ frontend/src/services/vat.js
@@ -1,0 +1,7 @@
+function vat2nif(vat) {
+  if (!vat) return null
+  if (vat.slice(0, 2) === 'ES') return vat.slice(2)
+  return vat
+}
+
+export { vat2nif }